requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
alembic==1.12.1
email-validator==2.1.0
aiofiles==23.2.1
Pillow==10.1.0
httpx==0.25.2
pyjwt==2.8.0
websockets==12.0
python-dotenv==1.0.0
geopy==2.4.1
